Task 2
Initial code. Takes a kml file that must have the correct name and placed in the exact directory then converts it to a shape file with a fixed name.
Minimal to no explanations. Program works with a pass code.
password is 555666
No self error checking. No conformation if program works other than a flash of data on the screen.